+++ genairics/scripts/gax_pipex_hello_world.py
@@ -7,6 +7,9 @@ within the environment running the script.
 
 The task needs to inherits/requires from genairics
 setupProject task as a bare minimum to setup logging.
+
+On the cluster server, your .bashrc should specify
+GAX_PIPEX and a PYTHONPATH that includes this scripts location.
 """
 import luigi
 #from luigi.util import requires
@@ -16,19 +19,19 @@ from genairics import setupProject, ProjectTask
 class HelloWorldTask(ProjectTask):
     name = luigi.Parameter(description="person's name")
     age = luigi.IntParameter(default=0, description="person's age")
+    loglevel = luigi.IntParameter(default=20, description="log level. 20 -> info; 30 -> warn (includes stdout)")
     
     def run(self):
-        from io import StringIO
-
-        logger = self.getLogger()
-        stdout = StringIO()
+        print = self.getPrint()
         print(
             'Hello', self.name,'!',
             *(('Are you really already',self.age,'?') if self.age else ()),
-            file = stdout
+            level = self.loglevel
         )
-        logger.info(stdout.getvalue())
-        print(stdout.getvalue(),end='')
+        #self.touchCheckpoint()
+
+    def output(self):
+        return self.CheckpointTarget()
 
 # Set this to the name of the class that will run as the pipeline
 gax_pipex_name = 'HelloWorldTask'
@@ -39,11 +42,12 @@ if __name__ == '__main__':
 
     moduleName = os.path.basename(__file__)[:-3]
     with local.env(
-            GAX_PIPEX = "{}.{}".format(moduleName, gax_pipex_name),
-            PYTHONPATH = ".:{}".format(os.getenv('PYTHONPATH'))
+            GAX_PIPEX = "{}.{}".format(moduleName, gax_pipex_name), #when submitting to queue will not yet work
+            PYTHONPATH = "{}:{}".format(os.path.dirname(__file__),os.getenv('PYTHONPATH')) #as these 2 vars are not passed
     ):
-        stdout = local['genairics'](*sys.argv[1:])
+        rc,stdout,stderr = local['genairics'][sys.argv[1:]].run()
         if stdout: print(stdout,end='')
+        if stderr: print(stderr,end='') #logging output to stderr and logfile
 
     # Example making genairics task to pipeline
     # with local.env(GAX_PIPEX='genairics.datasources.BaseSpaceSource'):